golint; move http scanner to its own module; allow partial integration tests to be run with TEST_MODULES;
