+++ integration_tests/http/setup.sh
@@ -11,11 +11,11 @@ if docker ps --filter "name=$CONTAINER_NAME" | grep $CONTAINER_NAME; then
 fi
 
 # First attempt to just launch the container
-if ! docker run --rm --name $CONTAINER_NAME -itd $CONTAINER_TAG; then
+if ! docker run --rm --name $CONTAINER_NAME -td $CONTAINER_TAG; then
     # If it fails, build it from ./container/Dockerfile
     docker build -t $CONTAINER_TAG ./container
     # Try again
-    docker run --rm --name $CONTAINER_NAME -itd $CONTAINER_TAG
+    docker run --rm --name $CONTAINER_NAME -td $CONTAINER_TAG
 fi
 
 echo -n "http/setup: Waiting on $CONTAINER_NAME to start..."
